requirements.txt
# =====================================================
# VISION EDGE HEALTHCARE - PYTHON REQUIREMENTS  
# PostgreSQL + Redis + FastAPI WebSocket Stack + Supabase Realtime
# =====================================================

# FastAPI and WebSocket
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3

# Database drivers
asyncpg==0.29.0          # PostgreSQL async driver
redis[hiredis]==5.0.1    # Redis async client with performance boost

# Supabase integration for realtime events
supabase>=2.0.0

# Data validation and serialization
pydantic==2.5.0

# HTTP client (for testing)
httpx==0.25.2
requests==2.31.0

# JSON handling
orjson==3.9.10           # Fast JSON serialization

# Core dependencies
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0

# Supabase integration
supabase>=2.0.0
python-dotenv>=1.0.0

# Date/Time utilities
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Security
cryptography>=41.0.0

# Development tools (optional for production)
pytest>=7.4.0
pytest-asyncio==0.21.1

# Utilities
click==8.1.6
tqdm==4.65.0
colorama==0.4.6
python-dateutil==2.8.2
sshkeyboard>=2.3.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1
